api/pyproject.toml
[project]
name = "mobox"
version = "0.1.0"
description = "Run multiple different AI agents in isolated sandboxes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "pyyaml>=6.0.0",
    "modal>=0.73.0",
    "shared",
    "pytest>=8.0.0",
]

[tool.uv.sources]
shared = { path = "../shared" }
